move instance update funcs
